stivale2: add defines (read info)
Since stivale2 tells us that the `uint64_t flags` member in `stivale2_struct_tag_firmware` can be `0` for UEFI or `1` for BIOS, then I decided that it's a good idea to add defines for these two. In a nutshell, defines for `STIVALE2_FIRMWARE_UEFI` and `STIVALE2_FIRMWARE_BIOS` with values `0` and `1`, respectively, added to the file. Code style and naming style are followed.